Moving to a better architecture
