+++ src/client/assets/.DS_Store
