+++ src/client/assets/images/CraftParty.png
