+++ src/client/assets/images/drop-bear.png
