+++ src/client/assets/images/emsc-logo.png
