+++ src/client/assets/images/small-group-pnw.png
