+++ src/client/assets/images/three-people-chill-drinks.png
