+++ src/client/assets/images/whole-group-chambers.png
