+++ src/main.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import { Router } from "./routes/Router.tsx";
+import { Router } from "./client/routes/Router.tsx";
 import { MantineProvider } from "@mantine/core";
 import "@mantine/core/styles.css";
-import { theme } from "./config/mantineTheme.ts";
+import { theme } from "./client/config/mantineTheme.ts";
 import "@mantine/carousel/styles.css";
-import AppFooter from "./componenets/Footer.tsx";
+import AppFooter from "./client/componenets/Footer.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
