+++ src/server/config/firebase-config.js
@@ -1,0 +1,7 @@
+const admin = require("firebase-admin");
+
+const serviceAccount = require("./serviceAccount.json");
+
+admin.initializeApp({
+  credential: admin.credential.cert(serviceAccount)
+});
